Redesigne registrer sanger

+++ endreAlbumTemplate.php
@@ -16,28 +16,42 @@ $album=str_replace("artister/", "", $artistMedMappe);//Tar vekk mappevisningen, 
 
 print ($artist);
 ?>
+ <form name="sporForm" action="<?php print ($album);?>.php" method="POST" enctype="multipart/form-data">
+
+
+
 
 <h2> <a href="index.php">Tilbake</a></h2>
+<div class="container-spor">
 
+<div class="h2album">
 <h1>Dette albumet heter <?php print ($album);?></h1> <br>
-<img src="/bilder/<?php print ($album);?>.jpeg"  style='width:200px;height:200px;'>
+</div>
+
+<div class="albumCover">
+<img class="albumsidebilde" src="/bilder/<?php print ($album);?>.jpeg"  style='width:200px;height:auto;'>
+</div>
 
 
 
-<form name="sporForm" action="<?php print($album);?>.php" enctype='multipart/form-data'>
+   
 
-<?php include("functions.php"); genererSpor($album);?> <br> <!-- Genererer <input> for hvert spor -->
+        <?php include("functions.php"); genererSpor($album);?>  <!-- Genererer <input> for hvert spor -->
 
-</form>
+<!-- ----------------------------------------------------------------- -->
+<div class="knapper">
+        <input type="submit" name="submit" value="LAGRE ENDRINGER">
+        <input type="submit" name="nyttSpor" value="NYTT SPOR">
+        <input type="submit" name="slettSpor" value="SLETT SISTE SPOR">
+</div>
+</div>
+    </form>
+
+<br>
 
 
 
-<form action="<?php print($album);?>.php" method="POST">
-    <input type="submit" name="nyttSpor" value="NYTT SPOR">
-    <input type="submit" name="slettSpor" value="SLETT SISTE SPOR">
-    <input type="submit" name="submit" value="LAGRE ENDRINGER">
 
-</form>
 
 
 
@@ -83,6 +97,10 @@ if (isset($_POST["slettSpor"])){
         $radAntallSpor=mysqli_fetch_array($resultatAntallSpor);
         $antallSpor=$radAntallSpor["AntallSpor"];
 
+if ($antallSpor>0){ //passer på at databasen ikke kan vise negative verdier på antall spor
+    $lydfil="lydfiler/spor" . $antallSpor . "album" . $globalAlbumNr . ".wav";
+    unlink($lydfil);
+
     //Sletter sporet
     $sqlSlett="DELETE FROM Spor WHERE Spor.SporNr='$antallSpor' AND Spor.AlbumNR='$globalAlbumNr';";
     mysqli_query($db, $sqlSlett);
@@ -90,19 +108,70 @@ if (isset($_POST["slettSpor"])){
     //Reduserer antall spor med 1 
     $sqlSETned="UPDATE Album SET AntallSpor = AntallSpor - 1 WHERE AlbumNavn='$album';";
     mysqli_query($db, $sqlSETned);
+}
 
 print("<meta http-equiv='refresh' content='0;url=$album.php'>"); //returnerer til albumsiden
 //--------------------------------------------------------------------------------------------------------
 
-
+}
 
 
 
 //----------------------------------------------------------------------------------------------------------
 //Lagrer endringer gjort på spor
+if (isset($_POST["submit"])){
+    include("dbTilkobling.php");
 
+        $sqlSELECTendring="SELECT * FROM Spor WHERE Spor.AlbumNr='$globalAlbumNr';";
+        $resultatEndring=mysqli_query($db, $sqlSELECTendring);
+        $antallRaderEndring=mysqli_num_rows($resultatEndring);
+
+        $lengdeKalkulert=[];
+        $sqlENDRE="";
+
+
+        for ($r=1;$r<=$antallRaderEndring;$r++){
+            $sportittel=$_POST["tittel$r"];
+            $sporlengde=$_POST["lengde$r"];
+            $sporisrc=$_POST["isrc$r"];
+
+            $filnavn=$_FILES ["fil$r"]["name"];  // filnavn på opplastet fil  
+            $filtype=$_FILES ["fil$r"]["type"];  // filtype på opplastet fil 
+            $filstorrelse=$_FILES ["fil$r"]["size"];  // filstørrelse på opplastet fil  
+            $tmpnavn=$_FILES ["fil$r"]["tmp_name"];    // midlertidig navn på opplastet fil 
+            $nyttnavn="lydfiler/"."spor" .$r ."album" . $globalAlbumNr ."." . "wav";  // mappe- og filnavn på opplastet fil 
+
+            //---------Regner ut hvor lang filen er og laster det opp til databasen
+            if ($filstorrelse>0){
+                $sekunder=round($filstorrelse/288000);
+                $min=round($sekunder/60);
+                $sekunderPluss=$sekunder-($min*60);
+                $tidTotalt=$min . ":" . $sekunderPluss;
+                array_push($lengdeKalkulert, $tidTotalt);
+                $sqlENDRE .="UPDATE Spor SET Lengde = '$tidTotalt' WHERE Spor.AlbumNr='$globalAlbumNr' AND Spor.SporNr='$r';";
+
+                move_uploaded_file($tmpnavn,$nyttnavn);
+            }
+        }
+
+        for ($r=1;$r<=$antallRaderEndring;$r++){
+            $l=$r-1;
+            $sportittel=$_POST["tittel$r"];
+            $sporlengde=$lengdeKalkulert[$l];
+            $sporisrc=$_POST["isrc$r"];
+
+            $sqlENDRE .="UPDATE Spor SET SporNavn = '$sportittel', ISRC = '$sporisrc' WHERE Spor.AlbumNr='$globalAlbumNr' AND Spor.SporNr='$r';";
+        }
+        mysqli_multi_query($db, $sqlENDRE);
+
+print("<meta http-equiv='refresh' content='0;url=$album.php'>"); //returnerer til albumsiden
 
 }
+
+//------------------------------------------------------------------------------------------------------------
+
+
+
 ?>
